% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeQI.R
\name{makeQI}
\alias{makeQI}
\title{Make a histogram, either plot or its data}
\usage{
makeQI(df, var, RHF, erMann, start_dato, slutt_dato)
}
\arguments{
\item{df}{dataframe from which output is to be made}

\item{var}{string defining which varable in the data frame to use}

\item{bins}{numeric vector defining the number of equally large groups}

\item{makeTable}{Logical that if TRUE function will return a data frame
containin the bin borders and count within each bin}
}
\value{
a graphical object or data frame
}
\description{
Short demo on how to produce dynamic content in a shiny app at Rapporteket
}
