% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_samlerapport.R
\name{samlerapport_server}
\alias{samlerapport_server}
\title{Server logic for samlerapport}
\usage{
samlerapport_server(
  id,
  enhetsvalg = c("Ahus", "Arendal", "Bodø", "Bærum", "Diakonhjemmet", "Drammen",
    "Elverum", "Flekkefjord", "Førde", "Gjøvik", "Hamar", "Hammerfest", "Haraldsplass",
    "Harstad", "Haugesund", "Haukeland", "Kalnes", "Kirkenes", "Kongsberg",
    "Kongsvinger", "Kristiansand", "Kristiansund", "Levanger", "Lillehammer", "Lofoten",
    "Lovisenberg", "Lærdal", "Mo i Rana", "Molde", "Mosjøen", "Namsos", "Narvik",
    "Nordfjord", "Notodden", "Odda", "Orkdal", "Ringerike", "Sandnessjøen", "Skien",
    "St. Olav", "Stavanger", "Stord", 
     "Tromsø", "Tynset", "Tønsberg", "Ullevål",
    "Vesterålen", "Volda", "Voss", "Ålesund")
)
}
\value{
A Shiny app server object
}
\description{
Server logic for samlerapport
}
