% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tertialrapport.R
\name{tertialrapport_server}
\alias{tertialrapport_server}
\title{Server logic for samlerapport}
\usage{
tertialrapport_server(
  id,
  enhetsvalg = c("Ahus", "Arendal", "Bodø", "Bærum", "Diakonhjemmet", "Drammen",
    "Elverum", "Flekkefjord", "Førde", "Gjøvik", "Hamar", "Hammerfest", "Haraldsplass",
    "Harstad", "Haugesund", "Haukeland", "Kalnes", "Kirkenes", "Kongsberg",
    "Kongsvinger", "Kristiansand", "Kristiansund", "Levanger", "Lillehammer", "Lofoten",
    "Lovisenberg", "Lærdal", "Mo i Rana", "Molde", "Mosjoen", "Namsos", "Narvik",
    "Nordfjord", "Notodden", "Odda", "Orkdal", "Rikshospitalet", "Ringerike",
    "Sandnessjøen", "Skien", "St. Olav", 
     "Stavanger", "Stord", "Tromsø",
    "Tynset", "Tønsberg", "Ullevål", "Vesterålen", "Volda", "Voss", "Ålesund")
)
}
\value{
A Shiny app server object
}
\description{
Server logic for samlerapport
}
