% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tertialrapport.R
\name{tertialrapport_server}
\alias{tertialrapport_server}
\title{Server logic for samlerapport}
\usage{
tertialrapport_server(
  id,
  enhetsvalg = c("Ahus", "Arendal", "Bodo", "Baerum", "Diakonhjemmet", "Drammen",
    "Elverum", "Flekkefjord", "Forde", "Gjovik", "Hamar", "Hammerfest", "Haraldsplass",
    "Harstad", "Haugesund", "Haukeland", "Kalnes", "Kirkenes", "Kongsberg",
    "Kongsvinger", "Kristiansand", "Kristiansund", "Levanger", "Lillehammer", "Lofoten",
    "Lovisenberg", "Laerdal", "Mo i Rana", "Molde", "Mosjoen", "Namsos", "Narvik",
    "Nordfjord", "Notodden", "Odda", "Orkdal", "Ringerike", "Sandnessjoen", "Skien",
    "St. Olav", "Stavanger", "Stord", 
     "Tromso", "Tynset", "Tonsberg", "Ullevaal",
    "Vesteraalen", "Volda", "Voss", "Aalesund")
)
}
\value{
A Shiny app server object
}
\description{
Server logic for samlerapport
}
