% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tertialrapport.R
\name{tertialrapport_server}
\alias{tertialrapport_server}
\title{Server logic for samlerapport}
\usage{
tertialrapport_server(
  id,
  enhetsvalg = c("Ahus", "Arendal", "Drammen", "Elverum", "Haukeland", "Rikshospitalet",
    "St. Olav")
)
}
\value{
A Shiny app server object
}
\description{
Server logic for samlerapport
}
